chore: refactor to use new component
