+++ packages/components/src/IconCountWithTooltip/IconCountWithTooltip.test.tsx
@@ -1,13 +1,16 @@
 import { render } from '../tests/utils'
 import { Default } from './IconCountWithTooltip.stories'
+
 import type { IconCountWithTooltipProps } from './IconCountWithTooltip'
 
 describe('IconCountWithTooltip', () => {
   it('validates the component behaviour', () => {
-    const { getByText } = render(
+    const { getByRole } = render(
       <Default {...(Default.args as IconCountWithTooltipProps)} />,
     )
 
-    expect(getByText('IconCountWithTooltip')).toBeInTheDocument()
+    const img = getByRole('img')
+
+    expect(img).toHaveAttribute('src', '/assets/icons/icon-comment.svg')
   })
 })
